src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-renderers-expected.txt

CONSOLE ERROR: line 83: Uncaught TypeError: host.createShadowRoot is not a function
